pyproject.toml
[tool.poetry]
name = "qml-mor"
version = "0.1.0"
description = "QML for Regression and Model Reduction"
authors = ["Mazen Ali <mazen.ali@itwm.fraunhofer.de>"]
readme = "README.md"
packages = [{include = "qml_mor"}]

[tool.poetry.dependencies]
python = "^3.10"
pythreejs = "^2.4.1"
ipywidgets = "7.7.1"
qiskit = {extras = ["visualization"], version = "^0.39.5"}
matplotlib = "^3.6.3"
scikit-learn = "^1.2.1"
pennylane = "^0.28.0"
torch = "^1.13.1"
pyyaml = "^6.0"
pennylane-lightning = {extras = ["gpu"], version = "^0.29.0"}


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"

[tool.mypy]
[[tool.mypy.overrides]]
module = "pennylane"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy.stats"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
